plugins/nx-cmake/src/executors/build/utils/buildProjectWithCMake/buildProjectWithCMake.ts
import type { BuildExecutorSchema } from '../../schema';
import { runCommand } from '@/command';
import { getCmake } from '../../../cmake/utils/getCmake/getCmake';
import { getCmakeBuildCommandArguments } from '../getCmakeBuildCommandArguments/getCmakeBuildCommandArguments';

export const buildProjectWithCMake = (
    workspaceRoot: string,
    projectRoot: string,
    options: BuildExecutorSchema,
): boolean => {
    const cmake = getCmake();
    const buildCommandArguments = getCmakeBuildCommandArguments(
        workspaceRoot,
        projectRoot,
        options,
    );
    const { success } = runCommand(cmake, ...buildCommandArguments);
    return success;
};
